add folders + middleware + fix README + lifespan + fix import + env
