+++ .env.example
@@ -1,2 +1,3 @@
 CLIENT_ID=local.66b0dfd36a6184.14182608
 CLIENT_SECRET=DR365b3V8sMtRRvKfdxzsddO1US3CNL
+# CLIENT_WEBHOOK=https://XXXXXX.bitrix24.ru/rest/1/tdww9m05xuclnc9n/
