+++ README.md
@@ -59,12 +59,14 @@ cp .env.example .env
 ```
 Убедитесь, что вы заполнили файл `.env` правильными значениями конфигурации перед запуском проекта.
 
+Учитывайте, что должны быть заполнены либо **CLIENT_ID** и **CLIENT_SECRET** для работы в режиме локального/тиражного приложения, либо **CLIENT_WEBHOOK** для работы с входящим в битрикс вебхуком.
+
 ### 6. Запустите проект
 
 Запустите проект:
 
 ```sh
-python main.py
+python run-server.py
 ```
 
 ## Примечания
