+++ run-server.py
@@ -1,5 +1,5 @@
-from examples.check_requests_fastapi import app
+from src.main import app
 import uvicorn
 
 if __name__ == "__main__":
-    uvicorn.run(app=app, host="127.0.0.1", port=8000)
+    uvicorn.run(app=app, host="127.0.0.1", port=8000, env_file=".env")
